Calm Spirit and Rebuke Spells
Change "calm spirit" to an untargeted AOE that attempts to stop combat in
the room, without the hitroll/damroll debuff. Add new "rebuke" spell to
spirit school that functions as the debuff part of "calm spirit" (modified
to be based on charisma instead of level, with a higher early effectiveness).